finalizing merge
